backend/requirements.txt
Django>=5.1
djangorestframework>=3.15
celery>=5.5
redis>=5.0
psycopg2-binary
django-storages
boto3
gunicorn
dj-rest-auth
django-allauth
requests
drf-spectacular
Pillow>=10.0
django-cors-headers
whitenoise
google-auth>=2.0
google-auth-oauthlib>=1.0
PyJWT>=2.0
cryptography>=41.0

